# careers-api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Careers API",
    "version": "1.0.0",
    "description": "API for managing career information"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "tags": [
    {
      "name": "Careers",
      "description": "Endpoints for managing career information"
    }
  ],
  "paths": {
    "/carriers": {
      "get": {
        "tags": ["Careers"],
        "summary": "Get all careers",
        "description": "Returns a list of all careers. Requires authentication and ADMIN or EMPLOYEE role.",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Career"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          }
        }
      },
      "post": {
        "tags": ["Careers"],
        "summary": "Create a new career",
        "description": "Creates a new career. Requires authentication and ADMIN or EMPLOYEE role.",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCareerRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Career created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Career"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          }
        }
      }
    },
    "/carriers/active": {
      "get": {
        "tags": ["Careers"],
        "summary": "Get active careers",
        "description": "Returns a list of all active careers. No authentication required.",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Career"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/carriers/{id}": {
      "get": {
        "tags": ["Careers"],
        "summary": "Get career by ID",
        "description": "Returns a single career by ID. Requires authentication and ADMIN or EMPLOYEE role.",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the career to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Career"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Career not found"
          }
        }
      },
      "patch": {
        "tags": ["Careers"],
        "summary": "Update a career",
        "description": "Updates an existing career. Requires authentication and ADMIN or EMPLOYEE role.",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the career to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCareerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Career updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Career"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID or input"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Career not found"
          }
        }
      },
      "delete": {
        "tags": ["Careers"],
        "summary": "Delete a career",
        "description": "Deletes a career by ID. Requires authentication and ADMIN or EMPLOYEE role.",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the career to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Career deleted successfully"
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Career not found"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Career": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Unique identifier for the career"
          },
          "name": {
            "type": "string",
            "description": "Name of the career"
          },
          "description": {
            "type": "string",
            "description": "Description of the career",
            "nullable": true
          },
          "isActive": {
            "type": "boolean",
            "description": "Indicates if the career is active",
            "default": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date and time when the career was created"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date and time when the career was last updated"
          }
        },
        "required": ["id", "name"]
      },
      "CreateCareerRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the career"
          },
          "description": {
            "type": "string",
            "description": "Description of the career"
          },
          "isActive": {
            "type": "boolean",
            "description": "Indicates if the career is active",
            "default": true
          }
        },
        "required": ["name"]
      },
      "UpdateCareerRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the career"
          },
          "description": {
            "type": "string",
            "description": "Description of the career"
          },
          "isActive": {
            "type": "boolean",
            "description": "Indicates if the career is active"
          }
        }
      }
    }
  }
}